backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
psycopg2-binary>=2.9
pydantic>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6
python-dateutil>=2.8
